requirements.txt
Flask 
openai 
# sqlite
google-generativeai
Pillow
# python dotenv
python-dotenv